La lentitud fue por un error, ya es mas rapido que celula

+++ src/gameoflife/HiloContador.java
@@ -1,20 +1,20 @@
 package gameoflife;
 
 public class HiloContador extends Thread{
-    private static final int MUERTA=0;
-    private static final int VIVA=1;
+    private static final byte MUERTA=0;
+    private static final byte VIVA=1;
     //S:supervivencia,N:Nacimiento
     private static final byte Smin=0;
     private static final byte Smax=1;
     private static final byte Nmin=2;
     private static final byte Nmax=3;
     
-    private final int[][] m,maux;
+    private final byte[][] m,maux;
     private int inicioX,finX,inicioY,finY;
     private volatile int vivas,muertas;
     private final int[] regla;
 
-    public HiloContador(int[][] m, int[][] maux, int inicioX, int finX, int inicioY, int finY,int[] regla) {
+    public HiloContador(byte[][] m, byte[][] maux, int inicioX, int finX, int inicioY, int finY,int[] regla) {
         this.m = m;
         this.maux = maux;
         this.inicioX = inicioX;
